Move mask conversion to separate file
